+++ cmd/webhook/app/webhook_test.go
@@ -19,7 +19,7 @@ package app
 import (
 	"testing"
 
-	"github.com/cert-manager/cert-manager/webhook-binary/app/options"
+	"github.com/cert-manager/cert-manager/pkg/webhook/options"
 )
 
 // Test to ensure flags take precedence over config options.
