+++ test/integration/LICENSES
@@ -12,7 +12,6 @@ github.com/cenkalti/backoff/v4,https://github.com/cenkalti/backoff/blob/v4.2.1/L
 github.com/cert-manager/cert-manager,https://github.com/cert-manager/cert-manager/blob/HEAD/LICENSE,Apache-2.0
 github.com/cert-manager/cert-manager/cmctl-binary,https://github.com/cert-manager/cert-manager/blob/HEAD/cmctl-binary/LICENSE,Apache-2.0
 github.com/cert-manager/cert-manager/integration-tests,https://github.com/cert-manager/cert-manager/blob/HEAD/integration-tests/LICENSE,Apache-2.0
-github.com/cert-manager/cert-manager/webhook-binary/app,https://github.com/cert-manager/cert-manager/blob/HEAD/webhook-binary/LICENSE,Apache-2.0
 github.com/cespare/xxhash/v2,https://github.com/cespare/xxhash/blob/v2.2.0/LICENSE.txt,MIT
 github.com/chai2010/gettext-go,https://github.com/chai2010/gettext-go/blob/v1.0.2/LICENSE,BSD-3-Clause
 github.com/containerd/containerd,https://github.com/containerd/containerd/blob/v1.7.0/LICENSE,Apache-2.0
